Hotfix (#8710)
* Run all tests in CI together (to be reverted)

* Changed celery_alt_dev_worker name from smoke_tests_worker -> alt_smoke_tests_worker

* Refactored stamping smoke tests
