+++ t/smoke/conftest.py
@@ -7,7 +7,7 @@ from pytest_docker_tools import container, fetch, network
 from t.smoke.workers.alt import *  # noqa
 from t.smoke.workers.dev import *  # noqa
 from t.smoke.workers.latest import *  # noqa
-from t.smoke.workers.legacy import *  # noqa
+from t.smoke.workers.other import *  # noqa
 
 
 @pytest.fixture
