+++ t/smoke/tests/stamping/workers/legacy.py
@@ -7,7 +7,7 @@ from pytest_docker_tools import build, container, fxtr
 from celery import Celery
 
 
-class CeleryLegacyWorkerContainer(CeleryWorkerContainer):
+class LegacyWorkerContainer(CeleryWorkerContainer):
     @property
     def client(self) -> Any:
         return self
@@ -22,18 +22,18 @@ class CeleryLegacyWorkerContainer(CeleryWorkerContainer):
 
     @classmethod
     def worker_name(cls) -> str:
-        return "celery4_tests_worker"
+        return "celery_legacy_tests_worker"
 
     @classmethod
     def worker_queue(cls) -> str:
-        return "celery4_tests_queue"
+        return "celery_legacy_tests_queue"
 
 
 celery_legacy_worker_image = build(
     path=".",
     dockerfile="t/smoke/workers/docker/pypi",
     tag="t/smoke/worker:legacy",
-    buildargs=CeleryLegacyWorkerContainer.buildargs(),
+    buildargs=LegacyWorkerContainer.buildargs(),
 )
 
 
@@ -42,14 +42,14 @@ celery_legacy_worker_container = container(
     environment=fxtr("default_worker_env"),
     network="{default_pytest_celery_network.name}",
     volumes={"{default_worker_volume.name}": defaults.DEFAULT_WORKER_VOLUME},
-    wrapper_class=CeleryLegacyWorkerContainer,
+    wrapper_class=LegacyWorkerContainer,
     timeout=defaults.DEFAULT_WORKER_CONTAINER_TIMEOUT,
 )
 
 
 @pytest.fixture
 def celery_legacy_worker(
-    celery_legacy_worker_container: CeleryLegacyWorkerContainer,
+    celery_legacy_worker_container: LegacyWorkerContainer,
     celery_setup_app: Celery,
 ) -> CeleryTestWorker:
     worker = CeleryTestWorker(celery_legacy_worker_container, app=celery_setup_app)
